+++ src/pylmrob/__init__.py
@@ -13,7 +13,14 @@ from pylmrob.anova import anova
 from pylmrob.bootstrap import bootstrap
 from pylmrob.control import Control
 from pylmrob.lmrob import LmRob, lmrob
-from pylmrob.rng import RState, r_sample_noreplace, r_set_seed, r_subsample_nonsingular
+from pylmrob.rng import (
+    RState,
+    r_norm_rand,
+    r_qnorm,
+    r_sample_noreplace,
+    r_set_seed,
+    r_subsample_nonsingular,
+)
 
 __all__ = [
     "Control",
@@ -24,6 +31,8 @@ __all__ = [
     "bootstrap",
     "lmrob",
     "make_generator",
+    "r_norm_rand",
+    "r_qnorm",
     "r_sample_noreplace",
     "r_set_seed",
     "r_subsample_nonsingular",